[FEATURE] Add AbstractHandlebarsController, define lib.handlebarsContent (JFBRELAU-1065)
`CTypeSwitchController`
=======================

A forward based on the `CType` of the current row is performed. The
name of the target controller is build according the following pattern:

```php
'JFB\\Userjungfrausite\\Controller\\' . ucfirst($cObj->data['CType']) . 'Controller'
```

If the controller is not present a "not implemented" messaged will be
returned (no exception).

`lib.handlebarsContent`
=======================

Declared in
`web/typo3conf/ext/userjungfrausite/Configuration/TypoScript/Plugin/Userjungfrausite/Setup.ts`.
Is used to invoke Extbase Bootstrap and run `CTypeSwitchController`.

Usage
-----

```
tt_content {
	connectionForm < lib.handlebarsContent
}
```

`renderHandlebarView`
=====================
This method is available in `AbstractController` in
EXT:userjungfrausite and can be used by extending from the abstract
class and invoking it in your Action.

Usage
-----

```
    public function listAction()
    {
        $data = [
            'activityCardList' => $this->getFormattedActivities()
        ];
        $this->renderHandlebarView('modules/activity_list/activity_list.hbs', $data);
    }
```